pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "google>=3.0.0",
    "google-genai>=1.67.0",
    "nltk==3.9.1",
    "numpy>=2.4.3",
    "pillow>=12.1.1",
    "sentence-transformers>=5.2.3",
]
